pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cfpb-connectors"
version = "0.1.0"
description = "LlamaIndex readers for loading CFBB data as Document objects"
readme = "README.md"
requires-python = ">=3.14.3"

dependencies = [
    "llama-index-core>=0.14.21",
    "requests>=2.33.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]

